added pagination class and pagination is ok

+++ delete.php
@@ -1,18 +1,22 @@
 <?php 
 	require_once 'core/init.php';
 
-	if(isset($_GET['id'])){
-		$id = $_GET['id'];
+	$user = new User();
+    if($user->isLoggedIn()) {
 
-		
-		$post = new Post();
-		try {
-	        $post->delete(array('id', '=', $id));
-	        // echo 'boom delete';
-	        header('location:index.php');
-	    } catch (Exception $e) {
-	        die($e->getMessage());
-	    }
+		if(isset($_GET['id'])){
+			$id = $_GET['id'];
+
+			$post = new Post();
+			try {
+		        $post->delete(array('id', '=', $id));
+		        //echo 'boom delete';
+		    	Redirect::to('index.php');
+		    } catch (Exception $e) {
+		        die($e->getMessage());
+		    }
+		}
+	} else {
+		Redirect::to(404);
 	}
-
 ?>
